build: add --size-only to S3 sync
While we're deploying multiple times a day, the data and static files
won't change as often. S3's sync command checks the file timestamp and
size by default, so files built in CI will always be changed and update
the last-modified header, triggering a redownload of cached files.

Passing the --size-only parameter overrides this behavior and only
updates a file in S3 if the sizes are different. This way, unchanged
info will be served from the cache even if deploys have run since the
last time. We're considering the risk of information changing without
changing the file size pretty low here, given the amount of data